削除　訂正

+++ app/controllers/items_controller.rb
@@ -1,5 +1,5 @@
 class ItemsController < ApplicationController
-  before_action :move_to_index, except: [:index, :show]
+  before_action :move_to_index, except: [:index, :show, :destroy]
   def index
     @items = Item.includes(:user).order('created_at DESC')
   end
@@ -23,6 +23,8 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    return unless item.user == current_user
+
     item.destroy
     redirect_to root_path
   end
